requirements.txt
requests==2.25.1
pyserial==3.5
